+++ pom.xml
@@ -41,10 +41,10 @@
 		</dependency>
 
 				<!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-security -->
-	<!--  <dependency>
+	  <dependency>
 		    <groupId>org.springframework.boot</groupId>
 		    <artifactId>spring-boot-starter-security</artifactId>
-		</dependency> -->	
+		</dependency>	
 		
 
 		<dependency>
