+++ src/main/java/com/EmployeeDataManagement/Controller/EmployeeLoginController.java
@@ -25,12 +25,13 @@ public class EmployeeLoginController {
 		String email=request.getParameter("email");
 		String password=request.getParameter("password");
 		EmployeeDataPojo employeeLoginDetails= employeeDataRepository.login(email, password);
-//		System.out.println(employeeLoginDetails);
+		System.out.println(employeeLoginDetails);
 //		System.out.println(employeeDataRepository.login(email, password));
 //		System.out.println(employeeLoginDetails.getName());
 		
 		if(employeeLoginDetails != null )
 		{
+			System.out.println(employeeLoginDetails);
 			HttpSession session=request.getSession();
 			String name=employeeLoginDetails.getName();
 			session.setAttribute("name",name);
@@ -39,8 +40,11 @@ public class EmployeeLoginController {
 		else
 		{
 			HttpSession session=request.getSession();
-			session.setAttribute("message","Invalid");
-			response.sendRedirect("login");
+			session.setAttribute("loginMessage","Invalid Email id or password");
+//			String message="Invalid Email id or password";
+//			Cookie ck=new Cookie("loginMessage",message);
+//			response.addCookie(ck);
+			response.sendRedirect("loginPage");
 		}
 	}
