+++ src/main/java/com/EmployeeDataManagement/Security/SecurityConfig.java
@@ -1,0 +1,22 @@
+package com.EmployeeDataManagement.Security;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig extends WebSecurityConfigurerAdapter  {
+	
+	@Override
+	 protected void configure(HttpSecurity http ) throws Exception
+	    {
+	     http.authorizeRequests()
+	     	.antMatchers("/").permitAll()
+//	     	.antMatchers("/login").permitAll()
+	     	.and().formLogin();
+	    }
+	
+	
+}
