+++ src/main/resources/application.properties
@@ -4,6 +4,7 @@ spring.mvc.view.suffix=.jsp
 spring.security.user.name=admin
 spring.security.user.password=admin
 
+
 spring.datasource.url=jdbc:mysql://localhost:3306/helloworld
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.username=root
